requirements.txt
matplotlib==3.7.1
Pillow==9.0.1
Pillow==10.1.0
streamlit==1.21.0
streamlit_drawable_canvas==0.9.2
torch==1.12.0
torchvision==0.13.0
